server/requirements.txt
fastapi==0.95.0
uvicorn[standard]
websockets==10.4
PyJWT==2.8.0
python-multipart==0.0.9